fix logging unit, use Duration

+++ tasks/msteams/src/main/java/com/walmartlabs/concord/plugins/msteams/TeamsV2Configuration.java
@@ -41,5 +41,5 @@ public interface TeamsV2Configuration {
      * Max duration to wait, in milliseconds, for retrying calls which failed due to
      * <code>429</code> response codes.
      */
-    int maxRetryWait();
+    long maxRetryWait();
 }
